+++ .vscode/settings.json
@@ -19,6 +19,9 @@
 		"array": "c",
 		"string": "c",
 		"string_view": "c",
-		"vector": "c"
+		"vector": "c",
+		"stdio.h": "c",
+		"term.h": "c",
+		"sstream": "c"
 	}
 }
