+++ Makefile
@@ -10,15 +10,25 @@ DFLAGS		=	-g3 -fsanitize=address
 LFTDIR		=	libft
 SRCDIR		=	src
 INCDIR		=	./include
-LIBS		=	-lncurses $(LFTDIR)/libft.a -lreadline -L /usr/local/opt/readline/lib
-INCS		=	-I $(LFTDIR) -I $(INCDIR) -I /usr/local/opt/readline/include
+LIBS		=	-lncurses $(LFTDIR)/libft.a -lreadline -L/opt/homebrew/opt/readline/lib
+INCS		=	-I $(LFTDIR) -I $(INCDIR) -I/opt/homebrew/opt/readline/include
 
 SRCS		=	$(SRCDIR)/core/main.c \
-				$(SRCDIR)/core/lexer/lexer.c \
-				$(SRCDIR)/core/lexer/deque.c \
-				$(SRCDIR)/core/lexer/free_struct.c \
-				$(SRCDIR)/core/lexer/token.c \
-				$(SRCDIR)/core/lexer/init_data.c \
+				$(SRCDIR)/core/parse/parse.c \
+				$(SRCDIR)/core/parse/print_parse_error.c \
+				$(SRCDIR)/core/parse/lexer/lexer.c \
+				$(SRCDIR)/core/parse/lexer/append_token.c \
+				$(SRCDIR)/core/parse/quote/quote.c \
+				$(SRCDIR)/core/parse/parser/parser.c \
+				$(SRCDIR)/core/parse/parser/parser_utils.c \
+				$(SRCDIR)/core/deque/deque.c \
+				$(SRCDIR)/core/free/free_struct.c \
+				$(SRCDIR)/core/free/free_tree.c \
+				$(SRCDIR)/core/token/token.c \
+				$(SRCDIR)/core/init/init_data.c \
+				$(SRCDIR)/core/init/init_tree.c \
+				$(SRCDIR)/core/init/init_lexer.c \
+				$(SRCDIR)/core/init/init_term.c \
 				$(SRCDIR)/core/envp/envp.c \
 				$(SRCDIR)/core/envp/env_utils.c \
 				$(SRCDIR)/core/env_parser/env_parser.c \
@@ -37,7 +47,10 @@ SRCS		=	$(SRCDIR)/core/main.c \
 				$(SRCDIR)/core/exec/free.c \
 				$(SRCDIR)/core/signal/signal.c \
 				$(SRCDIR)/core/exec/print_error.c \
-				$(SRCDIR)/core/builtin/export.c
+				$(SRCDIR)/core/builtin/export.c \
+				$(SRCDIR)/core/builtin/unset.c \
+				$(SRCDIR)/core/builtin/exit.c \
+				$(SRCDIR)/core/utils/utils.c
 OBJS		=	${SRCS:.c=.o}
 
 all :
