--- deque/dequeue.h
@@ -1,4 +1,0 @@
-#ifndef DEQUEUE_H
-# define DEQUEUE_H
-
-#endif
