+++ include/parser.h
@@ -125,4 +125,25 @@ void	ft_unset(t_simple_cmd *simple_cmd, t_deque *envs);
 void	ft_export(t_simple_cmd *simple_cmd, t_deque *d_envs);
 void	child_signal_handler(int signo);
 void	new_prompt_handler(int signo);
+int		env_name_check(char *name);
+void	print_parse_error(char *msg);
+void	print_syntax_error(char *msg);
+int		quote_check(char *line);
+int		parser(t_data *data);
+t_pipe	*init_pipe();
+t_cmd	*init_cmd();
+t_simple_cmd	*init_simple_cmd();
+t_redirects		*init_redirects();
+t_redirect		*init_redirect();
+int		get_pipe_size(t_pipe *pipe);
+void	set_built_in_flag(t_cmd *cmd, t_token *token);
+int		pipe_parser(t_pipe *pipe, t_deque *tokens);
+void 	open_token(t_lexer *lexer, t_data *data);
+void	append_token_content(t_lexer *lexer, t_data *data);
+void	close_token(t_lexer *lexer, t_data *data);
+void 	init_lexer(t_lexer *lexer);
+void 	append_element(t_lexer *lexer, t_data *data);
+void	init_tflag(struct termios *t);
+void	init_cc(struct termios *t);
+void	ft_exit(t_simple_cmd *simple_cmd);
 #endif
