+++ src/core/builtin/export.c
@@ -9,8 +9,8 @@ void	envs_sort(char **envs)
 	int		i;
 	int		j;
 
-	i = 0;
-	while (envs[i])
+	i = -1;
+	while (envs[++i])
 	{
 		j = i + 1;
 		while (envs[j])
@@ -27,7 +27,6 @@ void	envs_sort(char **envs)
 			free(next);
 			j++;
 		}
-		i++;
 	}
 }
 
@@ -47,8 +46,7 @@ void	print_export(t_deque *d_envs, int i)
 		ft_putstr_fd(name, 1);
 		if (content)
 		{
-			ft_putstr_fd("=", 1);
-			ft_putstr_fd("\"", 1);
+			ft_putstr_fd("=\"", 1);
 			ft_putstr_fd(content, 1);
 			ft_putstr_fd("\"", 1);
 		}
@@ -101,8 +99,8 @@ void	ft_export(t_simple_cmd *simple_cmd, t_deque *d_envs)
 	t_token	*temp;
 
 	flag = 0;
-	if (!simple_cmd)
-		return ;
+	// if (!simple_cmd)
+		// return ;
 	if (simple_cmd->argv->size == 1)
 		print_export(d_envs, 0);
 	else
@@ -114,32 +112,5 @@ void	ft_export(t_simple_cmd *simple_cmd, t_deque *d_envs)
 			temp = temp->next;
 		}
 	}
-	g_exit_status = flag;
-}
-
-void	ft_unset(t_simple_cmd *simple_cmd, t_deque *envs)
-{
-	char 	*name;
-	t_token *temp;
-	int		flag;
-
-	flag = 0;
-	if (simple_cmd->argv->size == 1 || simple_cmd->argv->size == 0)
-		return ;
-	temp = simple_cmd->argv->top->next;
-	while(temp)
-	{
-		name = get_env_name(temp->content);
-		if (env_name_check(name) == 1)
-			remove_env(envs, name);
-		if (env_name_check(name) == 0)
-		{
-			flag = 1;
-			print_error("unset", temp->content, 2);
-		}
-		free(name);
-		temp =temp->next;
-	}
-
 	g_exit_status = flag;
 }
