+++ src/core/deque/deque.c
@@ -25,7 +25,6 @@ void	append(t_deque *deque, t_token *token)
 		deque->bottom = token;
 	}
 	token->next = 0;
-	// deque->size = deque->size += 1;
 	deque->size += 1;
 }
 
@@ -44,7 +43,6 @@ void	appendleft(t_deque *deque, t_token *token)
 		deque->top = token;
 	}
 	token->pre = 0;
-	// deque->size = deque->size += 1;
 	deque->size += 1;
 }
 
@@ -67,7 +65,6 @@ t_token	*pop(t_deque *deque)
 	}
 	result->next = 0;
 	result->pre = 0;
-	// deque->size = deque->size -= 1;
 	deque->size -= 1;
 	return (result);
 }
@@ -91,7 +88,6 @@ t_token	*popleft(t_deque *deque)
 	}
 	result->next = 0;
 	result->pre = 0;
-	// deque->size = deque->size -= 1;
 	deque->size -= 1;
 	return (result);
 }
