+++ src/core/envp/envp.c
@@ -68,36 +68,4 @@ void	env_init(t_data *data, char **envp)
 		env_append(data->envs, envp[i]);
 		i++;
 	}
-	env_append(data->envs, "?=0");
-	///env_print_logic///
-	t_token *token = data->envs->top;
-	while (token)
-	{
-		printf("[%s] : [%s]\n", token->name, token->content);
-		token = token->next;
-	}
-	/////////////////
-	char **p = get_envs_pointer(data->envs);
-	int s = 0;
-	while (p[s])
-	{
-		ft_putstr_fd(p[s], 1);
-		write(1, "\n", 1);
-		s++;
-	}
-	free_envs_pointer(p);
-	env_append(data->envs, "??");
-	// env_append(data->envs, "??=");
-	// env_append(data->envs, "??=");
-	ft_putstr_fd(data->envs->bottom->name, 1);
-	write(1, "\n", 1);
-	p = get_envs_pointer(data->envs);
-	s = 0;
-	while (p[s])
-	{
-		ft_putstr_fd(p[s], 1);
-		write(1, "\n", 1);
-		s++;
-	}
-	free_envs_pointer(p);
 }
