+++ src/core/init/init_data.c
