+++ src/core/token/token.c
